Move influx interface into interfaces folder
