+++ app/Utils/InfluxDB/InfluxDBVersion1Wrapper.php
@@ -2,6 +2,7 @@
 
 namespace App\Utils\InfluxDB;
 
+use App\Interfaces\InfluxDBWrapperInterface;
 use App\Models\Speedtest;
 use Exception;
 use InfluxDB\Client;
